Migrate to extension.json
This just migrates the entrypoint into extension.json, and keeps a
backwards-compatible PHP entrypoint behind.

The intialization check to see whether Wikibase is installed is now
moved to the onBeforePageDisplay hook since that's what actually depends
upon Wikibase.

And the WIKIDATA_ORG_VERSION constant was removed since it wasn't used
anywhere outside of the credits, and double loading is not possible with
extension.json.

Change-Id: I29709dd603e796b3aba5b7f5dcfc683975ed2008