Do not use ./ in AutoloadNamespaces
AutoLoaderStructureTest::testPSR4Completeness in mediawiki/core ends up
failling because it can not handle a relative path ('./').  Though the
autoloader resolves it just fine.

To workaround the faulty test, uses an empty string to indicate the base
directory.  Note that ideally the php files should all be moved to ./src
or ./includes and AutoloadNamespaces set in consequence.

Test:

    $ php maintenance/shell.php
    Psy Shell v0.9.6 (PHP 7.0.27-0+deb9u1 — cli) by Justin Hileman
    >>> $a = new \WikidataOrg\Hooks();
    => WikidataOrg\Hooks {#322}
    >>>

Bug: T198077
Change-Id: I4cf627c9bdb81d5b45025b5f3998240f430fc30d