Add ApiMaxLagInfo hook
This hook will fetch the lag from the cached value.
If it is not there it will not alter maxlag.

A default configurable value of 3600 is applied.
This means that 10 hours of lag results in 10 seconds maxlag.

This should be enough to slow things down for now.

While writing this patch I realized that some of this code could
do with refactoring, but probably not worth worrying about that
right this second.

Bug: T221774
Change-Id: I99d35370dd9fae051bba608d4b68000dfb3f7872