Add "random item" messages for the sidebar
Currently on wikidata.org, the "randompage" message is overridden in
various languages to say "item" instead of "page". This copies the
approach used for the random lexeme link so that it can be translated
properly, and once the new messages are available, the sidebar can be
updated to use them.

Change-Id: I4582b549be90d7e492de5d55f6db1d11eb065eef