tests: Remove redundant WANObjectCache override
Since 2020 with change Ia60cb0bc18b23, the main cache is enabled
with a hash bag by default (and reset between each test), so this
override is identical to the default.

When an override is needed, the setMainCache() utility takes care of
everything automatically, including WANObjectCache, since that wraps
the main cache.

Change-Id: I4b3e1dc02043905f6561a9a3cb1662349f4544a6